Extract ClientCredentialsFlow
